Refactor delimiters helper file
