+++ tests/utils/delimitersHelper.test.ts
@@ -1,39 +1,39 @@
-import { findDelimiters } from '../../src/utils/stringHelpers';
+import { getStringDelimiters } from '../../src/utils/delimitersHelper';
 
-describe('String helpers', () => {
-  describe('findDelimiter ', () => {
+describe('Delimiters helper', () => {
+  describe('getStringDelimiters ', () => {
     it('should return "," as delimiter', () => {
-      const result = findDelimiters('1,2,3');
+      const result = getStringDelimiters('1,2,3');
       expect(result).toEqual([',']);
     });
 
     it('should return "\n" as delimiter', () => {
-      const result = findDelimiters('1\n2\n3');
+      const result = getStringDelimiters('1\n2\n3');
       expect(result).toEqual(['\n']);
     });
 
     it('should return ";" as delimiter', () => {
-      const result = findDelimiters('//;\n1;2\n');
+      const result = getStringDelimiters('//;\n1;2\n');
       expect(result).toEqual([';']);
     });
 
     it('should return "$" as delimiter', () => {
-      const result = findDelimiters('//$\n1$2\n');
+      const result = getStringDelimiters('//$\n1$2\n');
       expect(result).toEqual(['$']);
     });
 
     it('should return "***" as delimiter', () => {
-      const result = findDelimiters('//[***]\n1***2***3');
+      const result = getStringDelimiters('//[***]\n1***2***3');
       expect(result).toEqual(['***']);
     });
 
     it('should return "$$$$$" as delimiter', () => {
-      const result = findDelimiters('//[$$$$$]\n1$$$$$2$$$$$3');
+      const result = getStringDelimiters('//[$$$$$]\n1$$$$$2$$$$$3');
       expect(result).toEqual(['$$$$$']);
     });
 
     it('should return "*" and "%" as delimiters', () => {
-      const result = findDelimiters('//[*][%]\n1*2%3');
+      const result = getStringDelimiters('//[*][%]\n1*2%3');
       expect(result).toEqual(['*', '%']);
     });
   });
